+++ lib/cli.js
@@ -88,7 +88,7 @@ var methods = {
 
     // Process files
 
-    function parse(from) { return compiler.compile(sh.cat(from).replace(/^\uFEFF/g/* strips BOM */, ''), opt.compiler) }
+    function parse(from) { return compiler.compile(sh.cat(from).replace(/^\uFEFF/g, /* strips BOM */''), opt.compiler) }
     function toFile(from, to) { from.map(parse).join('\n').to(to[0]) }
     function toDir(from, to) { from.map(function(from, i) { parse(from).to(to[i]) }) }
     ;(opt.flow[1] == 'f' ? toFile : toDir)(from, to)
