+++ test/runner.html
@@ -10,10 +10,11 @@
     <script src="../node_modules/mocha/mocha.js"></script>
     <script src="../node_modules/expect.js/index.js"></script>
     <script>
-      mocha.setup('bdd');
+      mocha.setup('bdd')
     </script>
     <script src="../dist/riot/riot.js"></script>
     <script src="specs/tmpl.js"></script>
+    <script src="specs/observable.js"></script>
     <script src="runner.js"></script>
   </body>
 </html>
