+++ test/runner.js
@@ -16,6 +16,7 @@ if (isNode) {
   global.riot = require('../dist/riot/riot')
   global.expect = require('expect.js')
   require('./specs/tmpl')
+  require('./specs/observable')
 } else {
   mocha.run()
 }
